WIP: Refactoring KFServer and KFModel. * Removes concept of Seldon Protocol * Gives more context to KFModel to enable additional params in predict()/explain() * Cleaner class design for KFModel implementers with default passthrough for predict()/explain()
