+++ python/alibiexplainer/alibiexplainer/explainer.py
@@ -21,9 +21,7 @@ import numpy as np
 from alibiexplainer.anchor_images import AnchorImages
 from alibiexplainer.anchor_tabular import AnchorTabular
 from alibiexplainer.anchor_text import AnchorText
-from kfserving.protocols.seldon_http import SeldonRequestHandler
-from kfserving.protocols.tensorflow_http import TensorflowRequestHandler
-from kfserving.protocols.util import NumpyEncoder
+from kfserving.utils import NumpyEncoder
 from kfserving.server import Protocol, PREDICTOR_URL_FORMAT
 
 logging.basicConfig(level=kfserving.server.KFSERVER_LOGLEVEL)
@@ -44,16 +42,11 @@ class AlibiExplainer(kfserving.KFModel):
     def __init__(self,
                  name: str,
                  predictor_host: str,
-                 protocol: Protocol,
                  method: ExplainerMethod,
                  config: Mapping,
                  explainer: object = None):
         super().__init__(name)
-        self.protocol = protocol
-        if self.protocol == Protocol.tensorflow_http:
-            self.predict_url = PREDICTOR_URL_FORMAT.format(predictor_host, name)
-        else:
-            self.predict_url = SELDON_PREDICTOR_URL_FORMAT.format(predictor_host)
+        self.predict_url = PREDICTOR_URL_FORMAT.format(predictor_host, name)
         logging.info("Predict URL set to %s", self.predict_url)
         self.method = method
 
@@ -70,24 +63,18 @@ class AlibiExplainer(kfserving.KFModel):
         pass
 
     def _predict_fn(self, arr: Union[np.ndarray, List]) -> np.ndarray:
-        if self.protocol == Protocol.seldon_http:
-            resp = SeldonRequestHandler.predict(arr, self.predict_url)
-            return np.array(resp)
-        elif self.protocol == Protocol.tensorflow_http:
-            inputs = []
-            for req_data in arr:
-                if isinstance(req_data, np.ndarray):
-                    inputs.append(req_data.tolist())
-                else:
-                    inputs.append(str(req_data))
-            resp = TensorflowRequestHandler.predict(inputs, self.predict_url)
-            return np.array(resp)
-        else:
-            raise NotImplementedError
+        inputs = []
+        for req_data in arr:
+            if isinstance(req_data, np.ndarray):
+                inputs.append(req_data.tolist())
+            else:
+                inputs.append(str(req_data))
+        resp = self.predict(inputs, self.predict_url)
+        return np.array(resp)
 
-    def explain(self, inputs: List) -> Any:
+    def explain(self, request: Dict) -> Any:
         if self.method is ExplainerMethod.anchor_tabular or self.method is ExplainerMethod.anchor_images or self.method is ExplainerMethod.anchor_text:
-            explanation = self.wrapper.explain(inputs)
+            explanation = self.wrapper.explain(request["instances"])
             logging.info("Explanation: %s", explanation)
             return json.loads(json.dumps(explanation, cls=NumpyEncoder))
         else:
