+++ python/kfserving/kfserving/handlers/http/__init__.py
@@ -11,20 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from typing import Dict, List
-
-
-class RequestHandler(object):
-
-    def __init__(self, request: Dict):
-        self.request = request
-
-    def validate(self):
-        raise NotImplementedError
-
-    def extract_request(self) -> List:
-        raise NotImplementedError
-
-    def wrap_response(self, response: List) -> Dict:
-        raise NotImplementedError
