--- python/kfserving/kfserving/model.py
@@ -1,41 +1,0 @@
-# Copyright 2019 kubeflow.org.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from typing import List, Dict
-
-
-class KFModel(object):
-
-    def __init__(self, name: str):
-        self.name = name
-        self.ready = False
-
-    def load(self):
-        raise NotImplementedError
-
-    def preprocess(self, inputs: Dict) -> Dict:
-        return inputs
-
-    def predict(self, inputs: List) -> List:
-        raise NotImplementedError
-
-    def postprocess(self, inputs: List) -> List:
-        return inputs
-
-    def explain(self, inputs: List) -> Dict:
-        raise NotImplementedError
-
-    # TODO return type TBD
-    def detect_outlier(self, inputs: List):
-        raise NotImplementedError
