+++ python/kfserving/test/test_v1_inference_handler.py
@@ -16,7 +16,7 @@ import pytest
 import tornado
 import unittest.mock as mock
 import requests
-from kfserving.protocols.tensorflow_http import TensorflowRequestHandler
+from kfserving.protocols.v1_inference_handler import V1InferenceHandler
 
 
 def test_data_plane_tensorflow_protocol_client():
@@ -24,7 +24,7 @@ def test_data_plane_tensorflow_protocol_client():
     mock_resp.status_code = 200
     mock_resp.json = mock.Mock(return_value={"predictions": [[0.1, 0.9]]})
     with mock.patch('requests.post', return_value=mock_resp):
-        res = TensorflowRequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
+        res = V1RequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
         assert res == [[0.1, 0.9]]
 
 
@@ -32,7 +32,7 @@ def test_data_plane_tensorflow_protocol_client_raise_error():
     with mock.patch('requests.post', side_effect=
                     requests.exceptions.ConnectionError("fail to connect")):
         with pytest.raises(requests.exceptions.ConnectionError):
-            TensorflowRequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
+            V1RequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
 
 
 def test_data_plane_tensorflow_protocol_client_raise_http_error():
@@ -41,4 +41,4 @@ def test_data_plane_tensorflow_protocol_client_raise_http_error():
     mock_resp.reason = "Internal Server Error"
     with mock.patch('requests.post', return_value=mock_resp):
         with pytest.raises(tornado.web.HTTPError):
-            TensorflowRequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
+            V1RequestHandler.predict([[1, 2]], 'http://flower.default/v1/models/flower')
