+++ python/sklearnserver/sklearnserver/model.py
@@ -32,14 +32,14 @@ class SKLearnModel(kfserving.KFModel): #pylint:disable=c-extension-no-member
         self._joblib = joblib.load(model_file) #pylint:disable=attribute-defined-outside-init
         self.ready = True
 
-    def predict(self, body: List) -> List:
+    def predict(self, request: Dict) -> Dict:
         try:
-            inputs = np.array(body)
+            inputs = np.array(request["instances"])
         except Exception as e:
             raise Exception(
                 "Failed to initialize NumPy array from inputs: %s, %s" % (e, inputs))
         try:
             result = self._joblib.predict(inputs).tolist()
-            return result
+            return { "predictions" : result }
         except Exception as e:
             raise Exception("Failed to predict %s" % e)
