+++ python/sklearnserver/sklearnserver/test_model.py
@@ -32,4 +32,4 @@ def test_model():
     server.load()
     request = X[0:1].tolist()
     response = server.predict(request)
-    assert response == [0]
+    assert response["predictions"] == [0]
